requirements.txt
fastapi
uvicorn
transformers
torch
deep-translator
python-dotenv
google-search-results
sentencepiece
